+++ R/zzz.R
@@ -18,14 +18,15 @@
             utils::packageDescription("dtwSat")$Version) )
   
   ## Register TWDTW as a distance function into package proxy
-  pr_DB$set_entry(FUN   = .twdtwDist,
-                  names = c("TWDTW","twdtw"),
-                  loop  = FALSE,
-                  type  = "metric",
-                  description = "Time-Weighted Dynamic Time Warping",
-                  reference   = "Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016). A Time-Weighted Dynamic Time Warping method for land use and land cover mapping. IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9 (8), pp. 3729--3739. <doi: 10.1109/JSTARS.2016.2517118>."
-  );
+  is_there <- c("TWDTW","twdtw") %in% proxy::pr_DB$get_entry_names()
+  sapply(c("TWDTW","twdtw")[is_there], proxy::pr_DB$delete_entry)
   
+  proxy::pr_DB$set_entry(FUN   = .twdtwDist,
+                         names = c("TWDTW","twdtw"),
+                         loop  = FALSE,
+                         type  = "metric",
+                         description = "Time-Weighted Dynamic Time Warping",
+                         reference   = "Maus V, Camara G, Cartaxo R, Sanchez A, Ramos FM, de Queiroz GR (2016). A Time-Weighted Dynamic Time Warping method for land use and land cover mapping. IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, 9 (8), pp. 3729--3739. <doi: 10.1109/JSTARS.2016.2517118>."  )
 }
 
 #' @import zoo 
@@ -33,6 +34,7 @@
 #' @import ggplot2
 #' @import methods
 #' @import rgdal 
+#' @import snow 
 #' @importFrom proxy dist pr_DB
 #' @importFrom reshape2 melt
 #' @importFrom scales pretty_breaks date_format percent
@@ -45,6 +47,7 @@
 #' @importFrom lubridate month month<- day day<- year year<-
 #' @importFrom caret createDataPartition 
 #' @importFrom xtable xtable print.xtable
+#' @importFrom utils packageDescription flush.console 
 #' @useDynLib dtwSat, .registration = TRUE
 #' 
 NULL