+++ NEWS.md
@@ -1,3 +1,23 @@
+# dtwSat v0.2.4.9999
+
+* New features 
+
+  Include the function twdtwApplyParallel for TWDTW parallel processing using the package snow 
+  
+  Include writeRaster for twdtwRaster class 
+  
+  Improve tests and documentation 
+  
+  Improve memory usage of twdtwApply 
+  
+  Improve memory usage and speed of twdtwClassify 
+  
+* Fixing bugs
+  
+  Fix bug in twdtwAssess for class twdtwMatches 
+  
+  Fix bug in twdtwRaster 
+  
 # dtwSat v0.2.3
 
 * New features